refactor: threadMessage

+++ src/main/java/io/openmessaging/index/DichotomicIndex.java
@@ -106,7 +106,7 @@ public class DichotomicIndex {
     public void dumpInfo() {
         StringBuilder stringBuffer = new StringBuilder();
         for (int i = 0; i < size; i++) {
-            stringBuffer.append((int)tMeta.getLong(i << 3)).append(":").append(i).append(";");
+            stringBuffer.append(tMeta.getLong(i << 3)).append(":").append(i).append(";");
         }
         System.out.println(stringBuffer.toString());
     }
